# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Scholarship System API","description":"Backend service with JWT authentification","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Application Periods","description":"Admin endpoints for managing scholarship timelines"},{"name":"Decisions","description":"Administrative operations for scholarship assignment"},{"name":"Students","description":"Operations related to student profiles and data"},{"name":"Applications","description":"Management of scholarship applications"},{"name":"Authentication","description":"Endpoints for user registration and login using JWT"},{"name":"Reports","description":"Endpoints for generating and exporting scholarship reports"}],"paths":{"/api/students":{"get":{"tags":["Students"],"summary":"Get all students","description":"Returns a list of all registered students. Typically used by administrators.","operationId":"getAllStudents","responses":{"200":{"description":"Successfully retrieved list","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Student"}}}}}}},"post":{"tags":["Students"],"summary":"Create student profile","description":"Associates personal and academic data with the currently authenticated user. This must be done before applying for scholarships.","operationId":"createStudent","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudentRequestDTO"}}},"required":true},"responses":{"201":{"description":"Student profile successfully created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Student"}}}},"400":{"description":"Invalid input data or profile already exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Student"}}}}}}},"/api/scholarships/decide":{"post":{"tags":["Decisions"],"summary":"Run scholarship selection process","description":"Automatically assigns scholarships based on points and ranking. Status will be set to APPROVED for students within the limit, and REJECTED for others.","operationId":"decide","parameters":[{"name":"academicYear","in":"query","description":"The academic year for which to process decisions","required":true,"schema":{"type":"string"},"example":"2025-2026"},{"name":"limit","in":"query","description":"Maximum number of scholarships to be approved","required":true,"schema":{"type":"integer","format":"int32"},"example":50}],"responses":{"200":{"description":"Decision process completed successfully","content":{"*/*":{"schema":{"type":"string"}}}},"403":{"description":"Unauthorized - Only admins can run this process","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/auth/register":{"post":{"tags":["Authentication"],"summary":"Register a new user","description":"Creates a new user account (Student or Admin) and returns a JWT token.","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequestDTO"}}},"required":true},"responses":{"200":{"description":"User registered successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponseDTO"}}}},"400":{"description":"Email already exists or invalid data","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponseDTO"}}}}}}},"/api/auth/login":{"post":{"tags":["Authentication"],"summary":"User login","description":"Authenticates user credentials and returns a JWT token for further requests.","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequestDTO"}}},"required":true},"responses":{"200":{"description":"Login successful","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponseDTO"}}}},"401":{"description":"Invalid email or password","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponseDTO"}}}}}}},"/api/applications":{"post":{"tags":["Applications"],"summary":"Create a new application","description":"Creates a new scholarship application for the specified academic year for the currently logged-in student.","operationId":"createApplication","parameters":[{"name":"academicYear","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"201":{"description":"Application successfully created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Application"}}}},"400":{"description":"Invalid academic year or application already exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Application"}}}}}}},"/api/admin/application-periods":{"post":{"tags":["Application Periods"],"summary":"Create or Update an application period","description":"Defines when students can start and end their applications for a specific academic year.","operationId":"createPeriod","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PeriodRequestDTO"}}},"required":true},"responses":{"200":{"description":"Period successfully created/updated","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PeriodRequestDTO"}}}},"403":{"description":"Forbidden - Admin access required","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/reports/ranking/{academicYear}/pdf":{"get":{"tags":["Reports"],"summary":"Export ranking list as PDF","description":"Generates a PDF document containing the student ranking for a specific academic year. Restricted to Admin users.","operationId":"exportRankingPdf","parameters":[{"name":"academicYear","in":"path","description":"Academic year in format YYYY-YYYY","required":true,"schema":{"type":"string"},"example":"2025-2026"}],"responses":{"200":{"description":"PDF report successfully generated","content":{"application/pdf":{}}},"403":{"description":"Forbidden - Admin access required","content":{"*/*":{"schema":{"type":"string","format":"byte"}}}},"404":{"description":"Ranking for specified year not found","content":{"*/*":{"schema":{"type":"string","format":"byte"}}}}}}},"/api/applications/ranking/{academicYear}":{"get":{"tags":["Applications"],"summary":"Get scholarship ranking","description":"Returns a list of students ranked by their total points for a specific academic year.","operationId":"getRanking","parameters":[{"name":"academicYear","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully retrieved the ranking list","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ApplicationRankingDTO"}}}}}}}}},"components":{"schemas":{"StudentRequestDTO":{"type":"object","description":"Request object for submitting student's personal and academic information for scoring","properties":{"firstName":{"type":"string","description":"Student's first name","example":"Amer"},"lastName":{"type":"string","description":"Student's last name","example":"Hadić"},"email":{"type":"string","description":"Contact email address","example":"amer.hadic@student.unsa.ba"},"gpa":{"type":"number","format":"double","description":"Grade Point Average (GPA)","example":9.45,"maximum":10.0,"minimum":6.0},"incomePerMember":{"type":"number","format":"double","description":"Average monthly income per household member","example":450.0,"minimum":0},"socialStatus":{"type":"string","description":"Social status category for extra points","enum":["ORPHAN","SINGLE_PARENT","UNEMPLOYED_PARENTS","NORMAL"],"example":"WITHOUT_BOTH_PARENTS"},"locationType":{"type":"string","description":"Type of residence location","enum":["URBAN","RURAL"],"example":"URBAN"},"achievementType":{"type":"string","description":"Highest level of achievement reached","enum":["INTERNATIONAL","NATIONAL","REGIONAL","UNIVERSITY"],"example":"INTERNATIONAL"}},"required":["email","firstName","gpa","incomePerMember","lastName","locationType","socialStatus"]},"Student":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"gpa":{"type":"number","format":"double"},"incomePerMember":{"type":"number","format":"double"},"socialStatus":{"type":"string","enum":["ORPHAN","SINGLE_PARENT","UNEMPLOYED_PARENTS","NORMAL"]},"locationType":{"type":"string","enum":["URBAN","RURAL"]},"achievementType":{"type":"string","enum":["INTERNATIONAL","NATIONAL","REGIONAL","UNIVERSITY"]}}},"RegisterRequestDTO":{"type":"object","description":"Request object for new user registration","properties":{"email":{"type":"string","description":"User's unique email address","example":"novi.student@unsa.ba"},"password":{"type":"string","description":"User password (security policy: min 8 characters)","example":"Sifra123!","maxLength":2147483647,"minLength":8},"role":{"type":"string","description":"Role assigned to the user","enum":["STUDENT","ADMIN"],"example":"STUDENT"}},"required":["email","password"]},"AuthResponseDTO":{"type":"object","description":"Response object containing JWT token and basic user information","properties":{"token":{"type":"string","description":"JWT Bearer token for authenticating subsequent requests","example":"eyJhbGciOiJIUzI1NiJ9..."},"email":{"type":"string","description":"User's email address","example":"student@unsa.ba"},"role":{"type":"string","description":"User's assigned role in the system","enum":["ADMIN","STUDENT"],"example":"STUDENT"}}},"LoginRequestDTO":{"type":"object","description":"Request object for user authentication","properties":{"email":{"type":"string","description":"User's registered email address","example":"student@unsa.ba"},"password":{"type":"string","description":"User's password","example":"Password123!"}},"required":["email","password"]},"Application":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"student":{"$ref":"#/components/schemas/Student"},"academicYear":{"type":"string"},"totalPoints":{"type":"number","format":"double"},"status":{"type":"string","enum":["PENDING","APPROVED","REJECTED"]},"createdAt":{"type":"string","format":"date-time"}}},"PeriodRequestDTO":{"type":"object","description":"Request object for creating or updating a scholarship application period","properties":{"academicYear":{"type":"string","description":"The academic year identifier","example":"2025-2026"},"startDate":{"type":"string","format":"date","description":"The date when the application period opens","example":"2025-09-01"},"endDate":{"type":"string","format":"date","description":"The date when the application period closes","example":"2025-10-31"}}},"ApplicationRankingDTO":{"type":"object","description":"Data transfer object representing a student's position on the ranking list","properties":{"applicationId":{"type":"integer","format":"int64","description":"Unique ID of the application","example":101},"firstName":{"type":"string","description":"Student's first name","example":"Amar"},"lastName":{"type":"string","description":"Student's last name","example":"Hadić"},"totalPoints":{"type":"number","format":"double","description":"Calculated total points based on GPA, social status, and achievements","example":92.5},"rank":{"type":"integer","format":"int32","description":"Current rank on the list (1 being the highest)","example":1},"status":{"type":"string","description":"Status of the application (PENDING, APPROVED, REJECTED)","enum":["PENDING","APPROVED","REJECTED"],"example":"APPROVED"}}}},"securitySchemes":{"bearerAuth":{"type":"http","name":"bearerAuth","scheme":"bearer","bearerFormat":"JWT"}}}}